+++ versions.tf
@@ -1,10 +1,9 @@
 terraform {
-  experiments      = [module_variable_optional_attrs]
-  required_version = ">= 1.1, < 1.3"
+  required_version = ">= 1.3"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.0"
+      version = "~> 3.22"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"